requirements.txt
python-whois==0.9.4
validators==0.21.0
tldextract==3.1.0
lambda-warmer-py==0.6.0